requirements.txt
flask
python-dotenv
pyyaml
google-generativeai
markdown
gunicorn
